exporte members of krill-sequence to single files in krill-modules
