--- src/classes/krill-sequence.tsx
@@ -1,1321 +1,0 @@
-/**
- * penguins-eggs: krill_sequence
- *
- * author: Piero Proietti
- * mail: piero.proietti@gmail.com
- *
- */
-
-/**
- * Note: I'm using REACT here to get a TUI, 
- *       via ink library https://github.com/vadimdemedes/ink
- */
-
-
-/**
- * Ideally, I want to respect calamares way, remplementing the same (SEMPLIFIED) steps for CLI
- * 
- * Phase 2 - install.
- * 
- *  - partition:     partitions
- *  - mount:         mountFs, mountVfs
- *  - unpackfs:      unpackfs 
- *  - _restore:      eggs syncfrom --rootdir /tmp/calamares-krill-root/
- *  - sources-yolk:  execCalamaresModule('sources-yolk')
- *  - machineid:     machineid
- *  - fstab:         fstab
- *  - locale         locale
- *  - keyboard:      keyboard
- *  - localecfg:     localecfg
- *  - users:         users
- *  - displaymanager: autologin
- *  - networkcfg:    networkcfg
- *  - hwclock:
- *  - services-systemd:
- *  - bootloader-config: execCalamaresModule('bootloader-config')
- *  - grubcfg:       grubcfg
- *  - bootloader:    bootloaded
- *  - packages:      removeInstaller
- *  - luksbootkeyfile:
- *  - plymouthcfg;
- *  - initramfscfg:  initramfscfg
- *  - initramfs:     initramfs
- *  - removeuser:    removeuser
- *  - sources-yolk-unmount: execCalamaresModule('sources-yolk-unmount')
- *  - umount:     umountVfs, this.umountFs
- */
-
-import path from 'path'
-
-import { IRemix, IDistro, INet } from '../interfaces'
-import Settings from './settings'
-
-import React from 'react';
-import { render, RenderOptions } from 'ink'
-import Install from '../components/install'
-import Finished from '../components/finished'
-
-import fs from 'fs'
-import yaml from 'js-yaml'
-import shx from 'shelljs'
-import Utils from './utils'
-import cliAutologin = require('../lib/cli-autologin')
-import Pacman from './pacman';
-import { installer } from './incubation/installer'
-import Xdg from './xdg';
-import Distro from './distro'
-
-import { IInstaller, IDevices, IDevice } from '../interfaces'
-import { ICalamaresModule, ILocation, IKeyboard, IPartitions, IUsers } from '../interfaces/i-krill'
-import { exec } from '../lib/utils'
-
-import partition from '../lib/krill-modules/partitions'
-import bootloaderConfig from '../lib/krill-modules/bootloader-config'
-
-/**
- * hatching: installazione o cova!!!
- */
-export default class Sequence {
-
-   installer = {} as IInstaller
-
-   installTarget = '/tmp/calamares-krill-root'
-
-   verbose = false
-
-   echo = {}
-
-   efi = false
-
-   devices = {} as IDevices
-
-   users = {} as IUsers
-
-   network = {} as INet
-
-   partitions = {} as IPartitions
-
-   language = ''
-
-   region = ''
-
-   zone = ''
-
-   keyboardModel = ''
-
-   keyboardLayout = ''
-
-   keyboardVariant = ''
-
-   toNull = ' > /dev/null 2>&1'
-
-   settings = {} as Settings
-
-   remix = {} as IRemix
-
-   distro = {} as IDistro
-
-   luksName = ''
-
-   luksFile = ''
-
-   luksDevice = ''
-
-   luksMountpoint = ''
-
-   public partition = partition
-   public bootloaderConfig = bootloaderConfig
-
-
-   /**
-    * constructor
-    */
-   constructor(location: ILocation, keyboard: IKeyboard, partitions: IPartitions, users: IUsers, network: INet) {
-
-      this.installer = installer()
-      this.settings = new Settings()
-
-      this.language = location.language
-      this.region = location.region
-      this.zone = location.zone
-
-      this.keyboardModel = keyboard.keyboardModel
-      this.keyboardLayout = keyboard.keyboardLayout
-      this.keyboardVariant = keyboard.keyboardVariant
-
-      this.network = network
-
-      this.partitions = partitions
-
-      this.users = users
-
-      this.devices.efi = {} as IDevice
-      this.devices.boot = {} as IDevice
-      this.devices.root = {} as IDevice
-      this.devices.data = {} as IDevice
-      this.devices.swap = {} as IDevice
-
-      this.distro = new Distro(this.remix)
-
-      this.efi = fs.existsSync('/sys/firmware/efi/efivars')
-
-      // Per il restore dei dati
-      this.luksName = 'luks-eggs-backup'
-      this.luksFile = `/run/live/medium/live/${this.luksName}`
-      this.luksDevice = `/dev/mapper/${this.luksName}`
-      this.luksMountpoint = `/mnt`
-   }
-
-   /**
-    * install
-    * @param verbose 
-    * @param umount 
-    * @returns 
-    */
-   async install(verbose = false) {
-      this.verbose = verbose
-      this.echo = Utils.setEcho(this.verbose)
-      if (this.verbose) {
-         this.toNull = ''
-      }
-
-
-      // start
-      await this.settings.load()
-
-      // partition
-      let percent = 0.0
-      let message = ""
-      let isPartitioned = false
-
-      message = "Creating partitions"
-      percent = 0.01
-      try {
-         redraw(<Install message={message} percent={percent} />)
-         isPartitioned = await this.partition()
-      } catch (error) {
-         await Utils.pressKeyToExit(JSON.stringify(error))
-      }
-
-      if (isPartitioned) {
-
-         // formatting
-         message = "Formatting file system "
-         percent = 0.02
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            await this.mkfs()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // mountFs
-         message = "Mounting target file system "
-         percent = 0.03
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            await this.mountFs()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // mountVfs
-         message = "Mounting on target VFS "
-         percent = 0.06
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            await this.mountVfs()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // unpackfs
-         message = "Unpacking filesystem "
-         percent = 0.10
-         try {
-            //  spinner={true}
-            redraw(<Install message={message} percent={percent} />)
-            await this.unpackfs()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         /**
-          * IF RESTORE USERS DATA
-          */
-         if (fs.existsSync(this.luksFile)) {
-
-            // restoring users data
-            message = "Restore private data from backup "
-            percent = 0.37
-            let cmd = 'eggs syncfrom --rootdir /tmp/calamares-krill-root/'
-            try {
-               redraw(<Install message={message} percent={percent} spinner={true} />)
-               await exec(cmd, Utils.setEcho(true))
-            } catch (error) {
-               await Utils.pressKeyToExit(cmd)
-            }
-         }
-
-         // sources-yolk
-         if (this.distro.familyId === 'debian') {
-            message = 'sources-yolk'
-            percent = 0.40
-            try {
-               redraw(<Install message={message} percent={percent} spinner={true} />)
-               await this.execCalamaresModule('sources-yolk')
-            } catch (error) {
-               await Utils.pressKeyToExit(JSON.stringify(error))
-            }
-         }
-
-         // machineid
-         message = 'machineid'
-         percent = 0.41
-         try {
-            redraw(<Install message={message} percent={percent} spinner={true} />)
-            await this.machineId()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // fstab
-         message = "Creating fstab "
-         percent = 0.47
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            await this.fstab(this.partitions.installationDevice)
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // locale
-         message = "Locale "
-         percent = 0.47
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            await this.locale()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // keyboard
-         message = "settings keyboard "
-         percent = 0.48
-         try {
-            this.setKeyboard()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // localeCfg
-         message = "localeCfg"
-         percent = 0.50
-         try {
-            this.localeCfg()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-
-         // networkcfg
-         message = "networkcfg"
-         percent = 0.50
-         try {
-            this.networkcfg()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // hostname
-         message = "Create hostname "
-         percent = 0.53
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            await this.hostname()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // hosts
-         message = "Creating hosts "
-         percent = 0.60
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            await this.hosts()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // bootloader-config
-         message = "bootloader-config "
-         percent = 0.62
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            await this.bootloaderConfig()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // grubcfg
-         message = "grubcfg "
-         percent = 0.63
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            await this.grubcfg()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // bootloader
-         message = "bootloader "
-         percent = 0.64
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            await this.bootloader()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // initramfsCfg
-         message = "initramfs configure"
-         percent = 0.65
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            this.initramfsCfg(this.partitions.installationDevice)
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // initramfs
-         message = "initramfs "
-         percent = 0.67
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            await this.initramfs()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-
-         /**
-          * IF NOT RESTORE USERS DATA
-          */
-         if (!fs.existsSync(this.luksFile)) {
-            // delLiveUser
-            message = "Removing user live "
-            percent = 0.70
-            try {
-               redraw(<Install message={message} percent={percent} />)
-               await this.delLiveUser()
-            } catch (error) {
-               await Utils.pressKeyToExit(JSON.stringify(error))
-            }
-
-            // addUser
-            message = "Adding user "
-            percent = 0.73
-            try {
-               redraw(<Install message={message} percent={percent} />)
-               await this.addUser(this.users.name, this.users.password, this.users.fullname, '', '', '')
-            } catch (error) {
-               await Utils.pressKeyToExit(JSON.stringify(error))
-            }
-
-            // changePassword
-            message = "adding user password "
-            percent = 0.77
-            try {
-               redraw(<Install message={message} percent={percent} />)
-               await this.changePassword('root', this.users.rootPassword)
-            } catch (error) {
-               await Utils.pressKeyToExit(JSON.stringify(error))
-            }
-         }
-
-         // autologin
-         if (Pacman.isInstalledGui()) {
-            try {
-               message = "autologin GUI"
-               percent = 0.80
-               await Xdg.autologin(Utils.getPrimaryUser(), this.users.name, this.installTarget)
-               redraw(<Install message={message} percent={percent} />)
-            } catch (error) {
-               await Utils.pressKeyToExit(JSON.stringify(error))
-            }
-         } else { // autologin CLI remove DEFAULT
-            message = "autologin CLI"
-            percent = 0.80
-            try {
-               redraw(<Install message={message} percent={percent} />)
-               await cliAutologin.remove(this.installTarget)
-            } catch (error) {
-               await Utils.pressKeyToExit(JSON.stringify(error))
-            }
-         }
-
-         // cleanup
-         await cliAutologin.msgRemove(`${this.installTarget}/etc/motd`)
-         await cliAutologin.msgRemove(`${this.installTarget}/etc/issue`)
-
-         // removeInstaller
-         message = "remove installer"
-         percent = 0.87
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            await this.removeInstaller()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // sourcesYolkUmount
-         if (this.distro.familyId === 'debian') {
-            message = "sources yolk unmount"
-            percent = 0.92
-            try {
-               redraw(<Install message={message} percent={percent} />)
-               await this.execCalamaresModule('sources-yolk-unmount')
-            } catch (error) {
-               await Utils.pressKeyToExit(JSON.stringify(error))
-            }
-         }
-
-         // umountVfs
-         message = "umount VFS"
-         percent = 0.95
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            await this.umountVfs()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // umount
-         message = "umount"
-         percent = 0.97
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            await this.umountFs()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-
-         // finished
-         message = "finished"
-         percent = 100.0
-         try {
-            redraw(<Install message={message} percent={percent} />)
-            await this.finished()
-         } catch (error) {
-            await Utils.pressKeyToExit(JSON.stringify(error))
-         }
-      }
-   }
-
-   /**
-    * setTimezone
-    */
-   private async setTimezone() {
-      if (fs.existsSync('/etc/localtime')) {
-         const cmd = `chroot ${this.installTarget} unlink /etc/localtime`
-         await exec(cmd, this.echo)
-      }
-      const cmd = `chroot ${this.installTarget} ln -sf /usr/share/zoneinfo/${this.region}/${this.zone} /etc/localtime`
-      await exec(cmd, this.echo)
-   }
-
-   /**
-    *
-    * @param name
-    * @param password
-    * @param fullName
-    * @param roomNumber
-    * @param workPhone
-    * @param homePhone
-    */
-   private async addUser(name = 'live', password = 'evolution', fullName = '', roomNumber = '', workPhone = '', homePhone = ''): Promise<void> {
-
-      // Debian 
-      let cmd = `chroot ${this.installTarget} adduser ${name} --home /home/${name} --shell /bin/bash --disabled-password --gecos "${fullName},${roomNumber},${workPhone},${homePhone}" ${this.toNull}`
-      if (this.distro.familyId === 'archlinux') {
-         cmd = `chroot ${this.installTarget} useradd --create-home --shell /bin/bash ${name} ${this.toNull}`
-      }
-      await exec(cmd, this.echo)
-
-      cmd = `echo ${name}:${password} | chroot ${this.installTarget} chpasswd ${this.toNull}`
-      //  echo ${name}:${password} | chroot ${this.installTarget} chpasswd ${this.toNull}
-      await exec(cmd, this.echo)
-
-      // Debian
-      cmd = `chroot ${this.installTarget} usermod -aG sudo ${name} ${this.toNull}`
-      if (this.distro.familyId === 'archlinux') {
-         cmd = `chroot ${this.installTarget} usermod -aG wheel ${name}`
-      }
-
-      try {
-         await exec(cmd, this.echo)
-      } catch (error) {
-         await Utils.pressKeyToExit(cmd)
-      }
-   }
-
-   /**
-    * changePassword
-    * @param name
-    * @param newPassword
-    */
-   private async changePassword(name = 'live', newPassword = 'evolution') {
-      const cmd = `echo ${name}:${newPassword} | chroot ${this.installTarget} chpasswd ${this.toNull}`
-      await exec(cmd, this.echo)
-   }
-
-   /**
-    * delUser
-    * va corretto con users.conf di calamares
-    */
-   async delLiveUser() {
-      if (Utils.isLive()) {
-         const user: string = this.settings.config.user_opt
-
-         let userExists = false
-         try {
-            const cmd = `#!/bin/sh\ngetent passwd "${user}"  > /dev/null`
-            await exec(cmd, Utils.setEcho(this.verbose))
-            userExists = true
-         } catch (error) {
-            console.log(error)
-         } finally {
-            if (userExists) {
-               // debian family
-               let cmd = `chroot ${this.installTarget} deluser --remove-home ${user} ${this.toNull}`
-               if (this.distro.familyId === 'archlinux') {
-                  cmd = `chroot ${this.installTarget} sudo userdel -r ${user} ${this.toNull}`
-               }
-               await exec(cmd, this.echo)
-            }
-         }
-      }
-   }
-
-   /**
-    * grubInstall()
-    * @param target
-    * @param options
-    */
-   private async bootloader() {
-      await exec(`chroot ${this.installTarget} grub-install ${this.partitions.installationDevice} ${this.toNull}`, this.echo)
-      await exec(`chroot ${this.installTarget} update-grub ${this.toNull}`, this.echo)
-      await exec(`sleep 1 ${this.toNull}`, this.echo)
-   }
-
-   /**
-    * 
-   */
-   initramfsCfg(installDevice: string) {
-      if (this.distro.familyId === 'debian') {
-         // userSwapChoices = ['none', 'small', 'suspend', 'file']
-         const file = this.installTarget + '/etc/initramfs-tools/conf.d/resume'
-         let text = ''
-         if (this.partitions.userSwapChoice === 'none' || this.partitions.userSwapChoice === 'file') {
-            text += '#RESUME=none\n'
-         } else {
-            text += 'RESUME=UUID=' + Utils.uuid(this.devices.swap.name)
-         }
-         Utils.write(file, text)
-      } else if (this.distro.familyId === 'archlinux') {
-         console.log('initramfsCfg skipped')
-      }
-   }
-
-   /**
-    * initramfs()
-    */
-   private async initramfs() {
-      if (this.distro.familyId === 'debian') {
-         await exec(`chroot ${this.installTarget} mkinitramfs -o ~/initrd.img-$(uname -r) ${this.toNull}`, this.echo)
-         await exec(`chroot ${this.installTarget} mv ~/initrd.img-$(uname -r) /boot ${this.toNull}`, this.echo)
-      } else if (this.distro.familyId === 'archlinux') {
-         let initrdImg = Utils.initrdImg()
-         initrdImg = initrdImg.substring(initrdImg.lastIndexOf('/') + 1)
-         const cmd = `mkinitcpio -c ${path.resolve(__dirname, '../../mkinitcpio/manjaro/mkinitcpio-krill-inst.conf')} -g ${this.installTarget}/boot/${initrdImg}` // ${this.toNull}
-         try {
-            await exec(cmd, Utils.setEcho(true))
-         } catch (error) {
-            await Utils.pressKeyToExit(cmd)
-         }
-
-      }
-   }
-
-   /**
-    * fstab()
-    * @param devices
-    */
-   private async fstab(installDevice: string, crypted = false) {
-      let text = ''
-
-      /**
-       * crypttab
-       */
-      if (this.partitions.installationMode === 'full-encrypted') {
-         const crypttab = this.installTarget + '/etc/crypttab'
-         text = ``
-         text += `# /etc/crypttab: mappings for encrypted partitions.\n`
-         text += `#\n`
-         text += `# Each mapped device will be created in /dev/mapper, so your /etc/fstab\n`
-         text += `# should use the /dev/mapper/<name> paths for encrypted devices.\n`
-         text += `#\n`
-         text += `# See crypttab(5) for the supported syntax.\n`
-         text += `#\n`
-         text += `# NOTE: You need not list your root (/) partition here, but it must be set up\n`
-         text += `#       beforehand by the initramfs (/etc/mkinitcpio.conf). The same applies\n`
-         text += `#       to encrypted swap, which should be set up with mkinitcpio-openswap\n`
-         text += `#       for resume support.\n`
-         text += `#\n`
-         text += `# <name>               <device>                         <password> <options>\n`
-         text += `#swap_crypted was ${this.devices.swap.cryptedFrom}\n`
-         text += `swap_crypted UUID=${Utils.uuid(this.devices.swap.cryptedFrom)} none luks,discard\n`
-         text += `#root_crypted was ${this.devices.root.cryptedFrom}\n`
-         text += `root_crypted UUID=${Utils.uuid(this.devices.root.cryptedFrom)} none luks,swap\n`
-
-         Utils.write(crypttab, text)
-      }
-
-      const fstab = this.installTarget + '/etc/fstab'
-      let mountOptsRoot = ''
-      let mountOptsBoot = ''
-      let mountOptsData = ``
-      let mountOptsEfi = ''
-      let mountOptsSwap = ''
-
-      if (await this.isRotational(installDevice)) {
-         mountOptsRoot = 'defaults,relatime 0 1'
-         mountOptsBoot = 'defaults,relatime 0 1'
-         mountOptsData = 'defaults,relatime 0 1'
-         mountOptsEfi = 'defaults,relatime 0 2'
-         mountOptsSwap = 'defaults,relatime 0 2'
-      } else {
-         mountOptsRoot = 'defaults,noatime 0 1'
-         mountOptsBoot = 'defaults,noatime 0 1'
-         mountOptsData = 'defaults,noatime 0 1'
-         mountOptsEfi = 'defaults,noatime 0 2'
-         mountOptsSwap = 'defaults,noatime 0 2'
-      }
-
-      text = ``
-      text += `# ${this.devices.root.name} ${this.devices.root.mountPoint} ${this.devices.root.fsType} ${mountOptsRoot}\n`
-      text += `UUID=${Utils.uuid(this.devices.root.name)} ${this.devices.root.mountPoint} ${this.devices.root.fsType} ${mountOptsRoot}\n`
-
-      if (this.devices.boot.name !== `none`) {
-         text += `# ${this.devices.boot.name} ${this.devices.boot.mountPoint} ${this.devices.boot.fsType} ${mountOptsBoot}\n`
-         text += `UUID=${Utils.uuid(this.devices.boot.name)} ${this.devices.boot.mountPoint} ${this.devices.root.fsType} ${mountOptsBoot}\n`
-      }
-
-      if (this.devices.data.name !== `none`) {
-         text += `# ${this.devices.data.name} ${this.devices.data.mountPoint} ${this.devices.data.fsType} ${mountOptsData}\n`
-         text += `UUID=${Utils.uuid(this.devices.data.name)} ${this.devices.data.mountPoint} ${this.devices.data.fsType} ${mountOptsData}\n`
-      }
-
-      if (this.efi) {
-         text += `# ${this.devices.efi.name} ${this.devices.efi.mountPoint} vfat ${mountOptsEfi}\n`
-         text += `UUID=${Utils.uuid(this.devices.efi.name)} ${this.devices.efi.mountPoint} vfat ${mountOptsEfi}\n`
-      }
-      text += `# ${this.devices.swap.name} ${this.devices.swap.mountPoint} ${this.devices.swap.fsType} ${mountOptsSwap}\n`
-      text += `UUID=${Utils.uuid(this.devices.swap.name)} ${this.devices.swap.mountPoint} ${this.devices.swap.fsType} ${mountOptsSwap}\n`
-      Utils.write(fstab, text)
-   }
-
-   /**
-    * locale
-   */
-   private async locale() {
-      /**
-       * influcence: - /etc/default/locale
-       *             - /etc/locale.conf
-       *             - /etc/timezone
-       */
-      let defaultLocale = 'en_US.UTF-8'
-
-      // /etc/default/locale
-      let file = this.installTarget + '/etc/default/locale'
-      let content = ``
-      content += `LANG=${defaultLocale}\n`
-      content += `LC_CTYPE=${defaultLocale}\n`
-      content += `LC_NUMERIC=${defaultLocale}\n`
-      content += `LC_TIME=${defaultLocale}\n`
-      content += `LC_COLLATE=${defaultLocale}\n`
-      content += `LC_MONETARY=${defaultLocale}\n`
-      content += `LC_MESSAGES=${defaultLocale}\n`
-      content += `LC_PAPER=${defaultLocale}\n`
-      content += `LC_NAME=${defaultLocale}\n`
-      content += `LC_ADDRESS=${defaultLocale}\n`
-      content += `LC_TELEPHONE=${defaultLocale}\n`
-      content += `LC_MEASUREMENT=${defaultLocale}\n`
-      content += `LC_IDENTIFICATION=${defaultLocale}\n`
-      content += `LC_ALL=${defaultLocale}\n`
-      Utils.write(file, content)
-
-      // /etc/locale.conf
-      file = this.installTarget + '/etc/locale.conf'
-      Utils.write(file, content)
-
-      // timezone
-      try {
-         await this.setTimezone()
-      } catch (error) {
-         await Utils.pressKeyToExit(JSON.stringify(error))
-      }
-   }
-
-
-   /**
-    * setKeyboard
-    */
-   private async setKeyboard() {
-      /**
-      * influence: - /etc/default/keyboard (x11)
-      *            - /etc/vconsole.conf (console) 
-      *            - /etc/X11/xorg.conf.d/00-keyboard.conf
-      * 
-      * Problem: Actually don't update /etc/default/keyboard (x11)s
-      *          /etc/vconsole.conf is update in installed syste,
-      */
-
-      // systemd as default
-      let cmd = `chroot ${this.installTarget} localectl set-keymap ${this.keyboardLayout} ${this.toNull}`
-      if (!Utils.isSystemd()) {
-         cmd = `chroot ${this.installTarget} setupcon ${this.toNull}`
-      }
-
-      try {
-         await exec(cmd, this.echo)
-      } catch (error) {
-         console.log(error)
-         Utils.pressKeyToExit(cmd, true)
-      }
-
-      /**
-       * this must to be not necessary but...
-       * 
-       * force change /etc/default/keyboard (x11)
-       * force change /etc/vconsole.conf
-       */
-      let content = '# KEYBOARD CONFIGURATION FILE\n\n'
-      content += '# Consult the keyboard(5) manual page.\n\n'
-      content += 'XKBMODEL="' + this.keyboardModel + '"\n'
-      content += 'XKBLAYOUT="' + this.keyboardLayout + '"\n'
-      content += 'XKBVARIANT="' + this.keyboardVariant + '"\n'
-      content += 'XKBOPTIONS=""\n'
-      content += '\n'
-      content += 'BACKSPACE="guess"\n'
-      Utils.write(this.installTarget + '/etc/default/keyboard', content)
-
-      content = 'KEYMAP=' +  + this.keyboardModel + '"\n'
-      content += 'FONT=\n'
-      content += 'FONT_MAP=\n'
-      Utils.write(this.installTarget + '/etc/vconsole.conf', content)
-   }
-
-   /**
-    * localeCfg
-    * Enable the configured locales (those set by the user on the
-    * user page) in /etc/locale.gen, if they are available in the
-    * target system.
-    */
-   async localeCfg() {
-      /** 
-       * influence: - locale-gen
-       */
-      let supporteds: string[] = []
-      if (this.distro.familyId === 'debian') {
-         // Format: en_US.UTF-8 UTF-8
-         supporteds = fs.readFileSync('/usr/share/i18n/SUPPORTED', 'utf-8').split('\n')
-      } else if (this.distro.familyId === 'archlinux') {
-         // shx.exec('localectl list-locales > /tmp/SUPPORTED') // with await exec don't work! 
-         const supportedsSource = fs.readFileSync('/etc/locale.gen', 'utf-8').split('\n')
-
-         // Original Format: #en_US.UTF-8 UTF-8  
-         for (let line of supportedsSource) {
-            if (line.substring(0, 2) !== "# ") { // se non è un commento
-               line = line.substring(1) // Rimuove #
-            }
-            supporteds.push(line)
-         }
-         // Format: en_US.UTF-8 UTF-8  
-      }
-
-      const localeGenSource = fs.readFileSync(`${this.installTarget}/etc/locale.gen`, 'utf-8').split('\n')
-      let localeGenDest = ''
-      const krillBookmark = '#   Locales enabled by krill\n'
-      for (let line of localeGenSource) {
-         if (line.includes(krillBookmark)) {
-            break
-         }
-         localeGenDest += line + '\n'
-      }
-
-      localeGenDest += '\n'
-      localeGenDest += krillBookmark
-
-      const locales = [this.language]
-      if (this.language !== 'en_US.UTF-8') {
-         locales.push('en_US.UTF-8')
-      }
-      for (const supported of supporteds) {
-         for (const locale of locales) {
-            if (supported.includes(locale)) {
-               localeGenDest += `${supported}\n`
-            }
-         }
-      }
-      fs.writeFileSync(`${this.installTarget}/etc/locale.gen`, localeGenDest)
-   }
-
-   /**
-    * networkcfg
-    * 
-    * we have: 
-    * - debian: /etc/network/interface
-    * - ubuntu: netplan
-    * - manjaro: ? // ip address add 192.168.61/24 + dev enp6s18
-    */
-   private async networkcfg() {
-      if (this.distro.familyId === 'debian') {
-         // if netplan, don't create entries in /etc/network/interfaces
-         if (!Pacman.packageIsInstalled('netplan.io')) {
-            const file = this.installTarget + '/etc/network/interfaces'
-            let content = '# created by eggs\n\n'
-            content += 'auto lo\n'
-            content += 'iface lo inet loopback\n\n'
-            content += 'auto ' + this.network.iface + '\n'
-            content += 'iface ' + this.network.iface + ' inet ' + this.network.addressType + '\n'
-            if (this.network.addressType !== 'dhcp') {
-               content += '    address ' + this.network.address + '\n'
-               content += '    netmask ' + this.network.netmask + '\n'
-               content += '    gateway ' + this.network.gateway + '\n'
-            }
-            Utils.write(file, content)
-         }
-      }
-
-      /**
-       * resolv.conf
-       */
-      if (this.network.addressType !== 'dhcp') {
-         const file = this.installTarget + '/etc/resolv.conf'
-         let content = '# created by eggs\n\n'
-         content += 'domain ' + this.network.domain + '\n'
-         for (const element of this.network.dns) {
-            content += 'nameserver ' + element + '\n'
-         }
-         Utils.write(file, content)
-      }
-   }
-
-   /**
-    * hostname
-    */
-   private async hostname() {
-      await exec(`echo ${this.installTarget + '/etc/hostname'} > ${this.users.hostname} `, this.echo)
-   }
-
-   /**
-    * hosts
-    */
-   private async hosts() {
-      const file = this.installTarget + '/etc/hosts'
-      let text = '127.0.0.1 localhost localhost.localdomain\n'
-      if (this.network.addressType === 'static') {
-         text += `${this.network.address} ${this.users.hostname} pvelocalhost\n`
-      } else {
-         text += `127.0.1.1 ${this.users.hostname} \n`
-      }
-      text += `# The following lines are desirable for IPv6 capable hosts\n`
-      text += `:: 1     ip6 - localhost ip6 - loopback\n`
-      text += `fe00:: 0 ip6 - localnet\n`
-      text += `ff00:: 0 ip6 - mcastprefix\n`
-      text += `ff02:: 1 ip6 - allnodes\n`
-      text += `ff02:: 2 ip6 - allrouters\n`
-      text += `ff02:: 3 ip6 - allhosts\n`
-      fs.writeFileSync(file, text)
-   }
-
-   /**
-    * removeInstaller
-    */
-   private async removeInstaller() {
-      let file = `${this.installTarget}/usr/bin/penguins-links-add.sh`
-      let lines = []
-      let content = ''
-      if (fs.existsSync(file)) {
-         lines = fs.readFileSync(file, { encoding: 'utf8', flag: 'r' }).split('\n')
-         for (let i = 0; i < lines.length; i++) {
-
-            if (lines[i]) {
-               if (lines[i].search('penguins-krill.desktop') !== -1) {
-                  lines[i] += '#'
-               }
-               if (lines[i].search('penguins-clinstaller.desktop') !== -1) {
-                  lines[i] += '#'
-               }
-               if (lines[i].search('install-debian.desktop') !== -1) {
-                  lines[i] += '#'
-               }
-               content += lines[i] + '\n'
-            }
-         }
-      }
-      fs.writeFileSync(file, content)
-   }
-
-   /**
-    * unpackfs
-    */
-   private async unpackfs(): Promise<void> {
-      const cmd = `unsquashfs -d ${this.installTarget} -f ${this.distro.mountpointSquashFs}`
-      const echoYes = Utils.setEcho(true)
-      await exec(cmd, echoYes)
-   }
-
-   /**
-    * syncfs NOT-USED MORE
-    */
-   private async syncfs(): Promise<void> {
-
-      let cmd = ''
-      let f = ''
-      f += ' --filter="- /cdrom/*"'
-      f += ' --filter="- /dev/*"'
-      f += ' --filter="- /home/' + this.settings.config.user_opt + '/*"' // esclude da rsync solo utente live
-      f += ' --filter="- /live"'
-      f += ' --filter="- /media/*"'
-      f += ' --filter="- /mnt/*"'
-      f += ' --filter="- /proc/*"'
-      f += ' --filter="- /swapfile"'
-      f += ' --filter="- /sys/*"'
-      f += ' --filter="- /tmp/*"'
-
-      // boot
-      f += ' --filter="- /boot/efi*"'
-      f += ' --filter="- /boot/grub/device.map"'
-      f += ' --filter="- /boot/grub/grub.cfg"'
-      f += ' --filter="- /boot/grub/menu.lst"'
-
-      // etc
-      f += ' --filter="- /etc/fstab.d/*"'
-      f += ' --filter="- /etc/mtab"'
-      f += ' --filter="- /etc/popularity-contest.conf"'
-      f += ' --filter="- /etc/PolicyKit/PolicyKit.conf"'
-
-      // var
-      f += ' --filter="- /var/lib/dbus/machine-id"'
-
-      // Added for newer version of live-config/live-boot
-      f += ' --filter="- /lib/live/image"'
-      f += ' --filter="- /lib/live/mount"'
-      f += ' --filter="- /lib/live/overlay"'
-      f += ' --filter="- /lib/live/rootfs"'
-
-      // Added for symlink /lib
-      f += ' --filter="- /usr/lib/live/image"'
-      f += ' --filter="- /usr/lib/live/mount"'
-      f += ' --filter="- /usr/lib/live/overlay"'
-      f += ' --filter="- /usr/lib/live/rootfs"'
-
-      f += ' --filter="- /run/*"'
-
-      //       --progress \
-
-      cmd = `\
-      rsync \
-      --archive \
-      --delete -before \
-      --delete -excluded \
-      --info=progress2 \
-      ${f} \
-      / ${this.installTarget} ${this.toNull}`
-
-      await exec(cmd.trim(), this.echo)
-   }
-
-
-   /**
-    * mkfs
-    */
-   private async mkfs(): Promise<boolean> {
-      const result = true
-
-      if (this.efi) {
-         await exec(`mkdosfs -F 32 -I ${this.devices.efi.name} ${this.toNull}`, this.echo)
-      }
-
-      if (this.devices.boot.name !== 'none') {
-         if (this.devices.boot.fsType === undefined) {
-            this.devices.boot.fsType = `ext2`
-            this.devices.boot.mountPoint = '/boot'
-         }
-         await exec(`mke2fs -Ft ${this.devices.boot.fsType} ${this.devices.boot.name} ${this.toNull}`, this.echo)
-      }
-
-      if (this.devices.root.name !== 'none') {
-         await exec(`mke2fs -Ft ${this.devices.root.fsType} ${this.devices.root.name} ${this.toNull}`, this.echo)
-      }
-
-      if (this.devices.data.name !== 'none') {
-         await exec(`mke2fs -Ft ${this.devices.data.fsType} ${this.devices.data.name} ${this.toNull}`, this.echo)
-      }
-
-      if (this.devices.swap.name !== 'none') {
-         await exec(`mkswap ${this.devices.swap.name} ${this.toNull}`, this.echo)
-      }
-      return result
-   }
-
-   /**
-    * 
-    * @param device 
-    */
-   async ifMountedDismount(device = '') {
-
-      if ((await exec(`findmnt -rno SOURCE ${device}`)).data.trim() === device) {
-         await exec(`umount ${device} ${this.toNull}`, this.echo)
-         await exec('sleep 1', this.echo)
-      }
-   }
-
-   /**
-    * mount
-    */
-   private async mountFs(): Promise<boolean> {
-
-      if (!fs.existsSync(this.installTarget)) {
-         await exec(`mkdir ${this.installTarget} ${this.toNull}`, this.echo)
-      }
-
-      // root
-      await exec(`mount ${this.devices.root.name} ${this.installTarget}${this.devices.root.mountPoint} ${this.toNull}`, this.echo)
-      await exec(`tune2fs -c 0 -i 0 ${this.devices.root.name} ${this.toNull}`, this.echo)
-      await exec(`rm -rf ${this.installTarget}/lost+found ${this.toNull}`, this.echo)
-
-      // boot
-      if (this.devices.boot.name !== `none`) {
-         await exec(`mkdir ${this.installTarget}/boot -p ${this.toNull}`, this.echo)
-         await exec(`mount ${this.devices.boot.name} ${this.installTarget}${this.devices.boot.mountPoint} ${this.toNull}`, this.echo)
-         await exec(`tune2fs -c 0 -i 0 ${this.devices.boot.name} ${this.toNull}`, this.echo)
-      }
-
-      // data
-      if (this.devices.data.name !== `none`) {
-         await exec(`mkdir ${this.installTarget}${this.devices.data.mountPoint} -p ${this.toNull}`, this.echo)
-         await exec(`mount ${this.devices.data.name} ${this.installTarget}${this.devices.data.mountPoint} ${this.toNull}`, this.echo)
-         await exec(`tune2fs -c 0 -i 0 ${this.devices.data.name} ${this.toNull}`, this.echo)
-      }
-
-      // efi
-      if (this.efi) {
-         if (!fs.existsSync(this.installTarget + this.devices.efi.mountPoint)) {
-            await exec(`mkdir ${this.installTarget}${this.devices.efi.mountPoint} -p ${this.toNull}`, this.echo)
-            await exec(`mount ${this.devices.efi.name} ${this.installTarget}${this.devices.efi.mountPoint} ${this.toNull}`, this.echo)
-         }
-      }
-      return true
-   }
-
-   /**
-    * umountFs
-    */
-   private async umountFs(): Promise<boolean> {
-      // efi
-      if (this.efi) {
-         await this.umount(this.devices.efi.name)
-      }
-
-      // data
-      if (this.devices.data.name !== `none`) {
-         await this.umount(this.devices.data.name)
-      }
-
-      // boot
-      if (this.devices.boot.name !== `none`) {
-         await this.umount(this.devices.boot.name)
-      }
-
-      // root
-      await this.umount(this.devices.root.name)
-
-      return true
-   }
-
-   /**
-   * mountvfs()
-   */
-   private async mountVfs() {
-
-      await exec(`mkdir ${this.installTarget}/dev ${this.toNull}`, this.echo)
-      await exec(`mkdir ${this.installTarget}/dev/pts ${this.toNull}`, this.echo)
-      await exec(`mkdir ${this.installTarget}/proc ${this.toNull}`, this.echo)
-      await exec(`mkdir ${this.installTarget}/sys ${this.toNull}`, this.echo)
-      await exec(`mkdir ${this.installTarget}/run ${this.toNull}`, this.echo)
-
-      await exec(`mount -o bind /dev ${this.installTarget}/dev ${this.toNull}`, this.echo)
-      await exec(`mount -o bind /dev/pts ${this.installTarget}/dev/pts ${this.toNull}`, this.echo)
-      await exec(`mount -o bind /proc ${this.installTarget}/proc ${this.toNull}`, this.echo)
-      await exec(`mount -o bind /sys ${this.installTarget}/sys ${this.toNull}`, this.echo)
-      if (this.efi) {
-         await exec(`mount -o bind /sys/firmware/efi/efivars ${this.installTarget}/sys/firmware/efi/efivars ${this.toNull}`, this.echo)
-      }
-      await exec(`mount -o bind /run ${this.installTarget}/run ${this.toNull}`, this.echo)
-   }
-
-   /**
-    * 
-    */
-   private async umountVfs() {
-      await this.umount(`${this.installTarget}/dev/pts`)
-      await this.umount(`${this.installTarget}/dev`)
-      await this.umount(`${this.installTarget}/proc`)
-      await this.umount(`${this.installTarget}/run`)
-      if (this.efi) {
-         await this.umount(`${this.installTarget}/sys/firmware/efi/efivars`)
-      }
-      await this.umount(`${this.installTarget}/sys`)
-   }
-
-
-   /**
-    * 
-    * @param mountpoint 
-    */
-   private async umount(mountPoint = '') {
-      let message = 'umount: ' + mountPoint
-      if (Utils.isMountpoint(mountPoint)) {
-         try {
-            await exec(`umount ${mountPoint} ${this.toNull}`, this.echo)
-            await exec('sleep 1', this.echo)
-         } catch (error) {
-            message += + mountPoint + JSON.stringify(error)
-            redraw(<Install message={message} percent={1} />)
-            await Utils.pressKeyToExit(message)
-         }
-      }
-   }
-
-
-
-
-   /**
-    * Return lvmPartname, lvmSwapSize, lvmRootSize
-    */
-   public async lvmPartInfo(installDevice = '/dev/sda'): Promise<[string, number, number, number]> {
-
-      // Partizione LVM
-      const lvmPartname = shx.exec(`fdisk ${installDevice} -l | grep LVM | awk '{print $1}' | cut -d "/" -f3`).stdout.trim()
-      const lvmByteSize = Number(shx.exec(`cat /proc/partitions | grep ${lvmPartname}| awk '{print $3}' | grep "[0-9]"`).stdout.trim())
-      const lvmSize = lvmByteSize / 1024
-
-      // La partizione di root viene posta ad 1/4 della partizione LVM, limite max 100 GB
-      const lvmSwapSize = 8192
-      let lvmRootSize = lvmSize / 8
-      if (lvmRootSize < 20480) {
-         lvmRootSize = 20480
-      }
-      const lvmDataSize = lvmSize - lvmRootSize - lvmSwapSize
-      return [lvmPartname, lvmSwapSize, lvmRootSize, lvmDataSize]
-   }
-
-   /**
-    * isRotational
-    * @param device
-    */
-   private async isRotational(device: string): Promise<boolean> {
-      device = device.substring(4)
-      let response: any
-      let retVal = false
-
-      response = shx.exec(`cat /sys/block/${device}/queue/rotational`, { silent: this.verbose }).stdout.trim()
-      if (response === '1') {
-         retVal = true
-      }
-      return retVal
-   }
-
-
-   /**
-    * 
-    */
-   async execCalamaresModule(name: string) {
-      const moduleName = this.installer.multiarchModules + name + '/module.desc'
-      if (fs.existsSync(moduleName)) {
-         const calamaresModule = yaml.load(fs.readFileSync(moduleName, 'utf8')) as ICalamaresModule
-         let command = calamaresModule.command
-         if (command !== '' || command !== undefined) {
-            command += this.toNull
-            await exec(command, this.echo)
-         }
-      }
-   }
-
-   /**
-    * grubcfg
-    * - open /etc/default/grub
-    * - find GRUB_CMDLINE_LINUX_DEFAULT=
-    * - replace with GRUB_CMDLINE_LINUX_DEFAULT=
-    * 's/GRUB_CMDLINE_LINUX_DEFAULT=.*$/GRUB_CMDLINE_LINUX_DEFAULT=/g'
-    */
-   grubcfg() {
-      let file = `${this.installTarget}/etc/default/grub`
-      let content = ''
-      const grubs = fs.readFileSync(file, 'utf-8').split('\n')
-      for (let i = 0; i < grubs.length; i++) {
-         if (grubs[i].includes('GRUB_CMDLINE_LINUX_DEFAULT=')) {
-            if (this.partitions.installationMode === 'full-encrypted') {
-               grubs[i] = `GRUB_CMDLINE_LINUX_DEFAULT="resume=UUID=${Utils.uuid(this.devices.swap.name)}"`
-            } else {
-               grubs[i] = `GRUB_CMDLINE_LINUX_DEFAULT="quiet splash resume=UUID=${Utils.uuid(this.devices.swap.name)}"`
-            }
-         }
-         content += grubs[i] + '\n'
-      }
-      fs.writeFileSync(file, content, 'utf-8')
-   }
-
-
-   /**
-    * On Ubuntu 
-    * /etc/machine-id must exist to be re-created
-    * https://unix.stackexchange.com/questions/402999/is-it-ok-to-change-etc-machine-id
-    */
-   async machineId() {
-      let file = `${this.installTarget}/etc/machine-id`
-      if (fs.existsSync(file)) {
-         await exec(`rm ${file}`, this.echo)
-      }
-      await exec(`touch ${file}`)
-   }
-
-   /**
-    * only show the result
-    */
-   async finished() {
-      redraw(<Finished installationDevice={this.partitions.installationDevice} hostName={this.users.hostname} userName={this.users.name} />)
-      Utils.pressKeyToExit('Press a key to reboot...')
-      shx.exec('reboot')
-   }
-}
-
-const ifaces: string[] = fs.readdirSync('/sys/class/net/')
-
-/**
- * Occorre farglierlo rigenerare a forza
- * anche quando NON cambiano i dati
- * forceUpdate
- */
-function redraw(elem: JSX.Element) {
-   let opt: RenderOptions = {}
-
-   opt.patchConsole = false
-   opt.debug = false
-
-   shx.exec('clear')
-   render(elem, opt)
-}
