+++ src/krill/modules/bootloader-config-arch.ts
@@ -1,0 +1,15 @@
+/**
+ * https://stackoverflow.com/questions/23876782/how-do-i-split-a-typescript-class-into-multiple-files
+ */
+
+ import Sequence from '../krill-sequence'
+ 
+ /**
+  * 
+  * @param this 
+  */
+ 
+ export default async function bootloaderConfigArch(this: Sequence) {
+    console.log('bootloader Arch to do!')
+}
+
